chore(starknet_l1_provider): move flow test into unit test
Looks like our test runner doesn't support running integration tests
that depend on the crate with `testing` feature, so we can only use
integration tests that don't need `testing` on the same crate :(

This is almost all a MOVE with only one required change: the
bootstrap_e2e has gone stale and is broken, so i added an `ignore`.
The reason is that it wasn't run in the CI since required-features
doesn't add features, it just doesn't run the test if the feature isn't
there :/

Will fix the test in an upcoming PR.
